[net] Move ConnectionType to its own file
